Use TsFeature from xtask for --health
